refactor: extract auth functions into module
