+++ tests/helpers.py
@@ -3,7 +3,8 @@ from email.utils import parsedate_to_datetime
 from typing import Any, Optional
 from flask.testing import FlaskClient
 
-from arlo_server.routes import create_organization, UserType
+from arlo_server.auth import UserType
+from arlo_server.routes import create_organization
 from arlo_server.models import db, AuditAdministration, User
 
 DEFAULT_USER_EMAIL = "admin@example.com"
