+++ tests/test_jurisdictions.py
@@ -5,7 +5,8 @@ from typing import List
 
 from helpers import post_json, compare_json, assert_is_date, create_org_and_admin
 from test_app import client
-from arlo_server.routes import create_election, UserType
+from arlo_server.auth import UserType
+from arlo_server.routes import create_election
 from arlo_server.models import Jurisdiction
 from bgcompute import bgcompute_update_election_jurisdictions_file
